# 49_Se7enSquare-main/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "List all users",
        "description": "Returns every user in the in-memory store. Accessible by any valid role.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of user records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          }
        },
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create a new user",
        "description": "Creates a user record. Requires x-role: admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "User creation payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "Only admin can create users"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get a single user by ID",
        "description": "Fetches one user record by their numeric ID.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The matching user record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update an existing user",
        "description": "Partially updates a user (all fields optional). Requires x-role: admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric user ID to update",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update (all optional)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "Only admin can update users"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete a user",
        "description": "Permanently deletes a user and cascades to their memberships, posts, and reports. Requires x-role: admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric user ID to delete",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "User 4 deleted"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only admin can delete users"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/reports": {
      "get": {
        "operationId": "ReportsController_findAll",
        "summary": "List all reports",
        "description": "Returns all reports. Only accessible by admin or moderator roles.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET requires admin or moderator. POST requires any valid role. PATCH status requires admin or moderator. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of report records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only admin or moderator can view reports"
          }
        },
        "tags": [
          "reports"
        ]
      },
      "post": {
        "operationId": "ReportsController_create",
        "summary": "Submit a report",
        "description": "Creates a new report (content or user report). Any authenticated role may submit a report. Status defaults to \"pending\".",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET requires admin or moderator. POST requires any valid role. PATCH status requires admin or moderator. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Report submission payload (reporterId, targetType, targetId, reason)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly submitted report with status: pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}/status": {
      "patch": {
        "operationId": "ReportsController_updateStatus",
        "summary": "Update report status",
        "description": "Advances a report through the status lifecycle: pending → reviewed → resolved. Requires x-role: admin or moderator.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET requires admin or moderator. POST requires any valid role. PATCH status requires admin or moderator. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric report ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New status value. Allowed: pending | reviewed | resolved",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The report with updated status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDto"
                }
              }
            }
          },
          "403": {
            "description": "Only admin or moderator can update report status"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}": {
      "delete": {
        "operationId": "ReportsController_remove",
        "summary": "Delete a report",
        "description": "Permanently removes a report record. Requires x-role: admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | moderator | user. GET requires admin or moderator. POST requires any valid role. PATCH status requires admin or moderator. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric report ID to delete",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Report 1 deleted"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only admin can delete reports"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Se7enSquare Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | moderator | user | gamer"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "player01"
          },
          "email": {
            "type": "string",
            "example": "player@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "moderator",
              "user",
              "gamer"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "example": "Casual and competitive gamer"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "newplayer"
          },
          "email": {
            "type": "string",
            "example": "newplayer@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "moderator",
              "user",
              "gamer"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "minLength": 5,
            "maxLength": 160,
            "example": "Loves MMORPG and FPS."
          }
        },
        "required": [
          "username",
          "email",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "newplayer"
          },
          "email": {
            "type": "string",
            "example": "newplayer@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "moderator",
              "user",
              "gamer"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "minLength": 5,
            "maxLength": 160,
            "example": "Loves MMORPG and FPS."
          }
        }
      },
...
      "ReportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "reporterId": {
            "type": "number",
            "example": 3
          },
          "targetType": {
            "type": "string",
            "enum": [
              "post",
              "user",
              "community"
            ],
            "example": "post"
          },
          "targetId": {
            "type": "number",
            "example": 1
          },
          "reason": {
            "type": "string",
            "example": "Potential harassment in comments"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "reviewed",
              "resolved"
            ],
            "example": "pending"
          }
        },
        "required": [
          "id",
          "reporterId",
          "targetType",
          "targetId",
          "reason",
          "status"
        ]
      },
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "reporterId": {
            "type": "number",
            "example": 3
          },
          "targetType": {
            "type": "string",
            "enum": [
              "post",
              "user",
              "community"
            ],
            "example": "post"
          },
          "targetId": {
            "type": "number",
            "example": 1
          },
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 200,
            "example": "Abusive language found in comments"
          }
        },
        "required": [
          "reporterId",
          "targetType",
          "targetId",
          "reason"
        ]
      },
      "UpdateReportStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "reviewed",
              "resolved"
            ],
            "example": "reviewed"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}